semantic_search_api/requirements.txt
fastapi
uvicorn
pgvector
psycopg2-binary
sentence-transformers
pydantic
dotenv
boto3
